Fix: Acomodando capa de servicios en cliente

+++ src/main/java/com/mycompany/bazar/service/ClienteService.java
@@ -22,13 +22,13 @@ public class ClienteService implements IClienteService {
 
     @Override
     public List<Cliente> getClientes() {
-          return clienteRepo.findAll();
+        return clienteRepo.findAll();
     }
 
     @Override
     public Cliente findCliente(Long id) {
-        Cliente cli =clienteRepo.findById(id).orElse(null);
-        return  cli;
+        Cliente cli = clienteRepo.findById(id).orElse(null);
+        return cli;
     }
 
     @Override
@@ -37,13 +37,30 @@ public class ClienteService implements IClienteService {
     }
 
     @Override
-    public void deleteCliente(Long id) {
+    public Boolean deleteCliente(Long id) {
+        Cliente clienteAEliminar = this.findCliente(id);
+        if (clienteAEliminar == null) {
+            return false;
+        }
         clienteRepo.deleteById(id);
+        return true;
     }
 
     @Override
-    public void editCliente(Cliente cli) {
-        this.saveCliente(cli);  // no existe el metodo edit  
+    public Cliente editCliente(Cliente cli) {
+
+        Cliente cliente = this.findCliente(cli.getIdCliente()); // busco al cliente
+
+        if (cliente == null) {
+            return cliente;
+        }
+        // si es el mismo id, lo edita si no avisa
+        cliente.setNombre(cli.getNombre());
+        cliente.setApellido(cli.getApellido());
+        cliente.setDni(cli.getDni());
+
+        this.saveCliente(cliente);  // no existe el metodo edit  
+        return cliente;
     }
 
 }
